Move PyIteratorWrapper to shared ffi module
This code is useful anywhere FFI functions are being used.
